remove TODOs in example config
The example instance is meant to be easy to understand, not necessarily reflective of how the program will actually work in production. For example, we're fine with the simulation being run every 15 seconds instead of every hour (which is more realistic) in the example because that makes it easier to check that it works.